Georchestra gn3 integration fixes

+++ core/src/main/java/org/fao/geonet/kernel/security/ldap/LDAPSynchronizerJob.java
@@ -190,7 +190,9 @@ public class LDAPSynchronizerJob extends QuartzJobBean {
                 return input.getId();
             }
         });
-        userGroupRepository.deleteAllByIdAttribute(UserGroupId_.userId, userIds);
+        if (!userIds.isEmpty()) {
+            userGroupRepository.deleteAllByIdAttribute(UserGroupId_.userId, userIds);
+        }
         for (User u : usersFound) {
             long nbOfUserRecord = metadataRepository.count(MetadataSpecs.isOwnedByUser(u.getId()));
             if (nbOfUserRecord > 0) {
